UI: Add SSE client implementation and configuration Backend: Update SSE handler to send connection confirmation message

+++ backend/memories/model.go
@@ -18,7 +18,7 @@ type Memory struct {
 	EmbeddingModelId *int64
 }
 
-func (m *Memory) CosineSimilarity(other providers.Embeddings) (float64, error) {
+func (m *Memory) CosineSimilarity(other providers.Embeddings) (float32, error) {
 	return m.Embedding.CosineSimilarity(other)
 }
